[Widgets] Implemented displayName attribute for each widget properties.
In js/widgets.js there are some widget properties have displayName
attribute, it use for custom the label, but it is not functional.

See List.divider, OrderedList.divider, Collapsible.content_theme
and Accordion.content_theme in js/widgets.js.

The patch implemented the BWdiget.getPropertyDisplayName() to
make the attribute work.